Cleanup PoC and rename module
