+++ foundation-models/openai-v1/src/test/resources/mappings/createResponse.json
