+++ logback.xml
@@ -12,8 +12,8 @@
 		<appender-ref ref="CONSOLE"/>
 	</root>
 
-	  <logger name="org.apache.hc.client5.http.wire" level="DEBUG"/>
-	  <logger name="org.apache.http.wire" level="DEBUG"/>
+	<!--  <logger name="org.apache.hc.client5.http.wire" level="DEBUG"/>-->
+	<!--  <logger name="org.apache.http.wire" level="DEBUG"/>-->
 	<logger name="com.sap.ai.sdk" level="DEBUG"/>
 	<!-- Disable org.eclipse.jetty logging -->
 	<logger name="org.eclipse.jetty" level="OFF"/>
