+++ sample-code/spring-app/pom.xml
@@ -79,7 +79,7 @@
     </dependency>
     <dependency>
       <groupId>com.sap.ai.sdk.foundationmodels</groupId>
-      <artifactId>openai-responses</artifactId>
+      <artifactId>openai-v1</artifactId>
     </dependency>
     <dependency>
       <groupId>com.openai</groupId>
