+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/services/OpenAiV1Service.java
@@ -6,13 +6,12 @@ import com.openai.client.OpenAIClient;
 import com.openai.core.http.QueryParams;
 import com.openai.core.http.StreamResponse;
 import com.openai.models.ChatModel;
-import com.openai.models.chat.completions.ChatCompletion;
 import com.openai.models.chat.completions.ChatCompletionChunk;
 import com.openai.models.chat.completions.ChatCompletionCreateParams;
 import com.openai.models.responses.Response;
 import com.openai.models.responses.ResponseCreateParams;
 import com.openai.models.responses.ResponseStreamEvent;
-import com.sap.ai.sdk.foundationmodels.openai.responses.AiCoreOpenAiClient;
+import com.sap.ai.sdk.foundationmodels.openai.v1.AiCoreOpenAiClient;
 import javax.annotation.Nonnull;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -49,23 +48,6 @@ public class OpenAiV1Service {
     val params =
         ResponseCreateParams.builder().input(input).model(ChatModel.GPT_5).store(false).build();
     return client.responses().createStreaming(params);
-  }
-
-  /**
-   * Create a chat completion using the Chat Completions API
-   *
-   * @param input the input text to send to the model
-   * @return the chat completion response from the Chat Completions API
-   */
-  @Nonnull
-  public ChatCompletion createChatCompletion(@Nonnull final String input) {
-    val params =
-        ChatCompletionCreateParams.builder()
-            .addUserMessage(input)
-            .model(ChatModel.GPT_5)
-            .additionalQueryParams(QueryParams.builder().put("api-version", "2023-05-15").build())
-            .build();
-    return client.chat().completions().create(params);
   }
 
   /**
