+++ sample-code/spring-app/src/test/java/com/sap/ai/sdk/app/controllers/OpenAiV1Test.java
@@ -6,6 +6,7 @@ import com.sap.ai.sdk.app.services.OpenAiV1Service;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 @Slf4j
@@ -26,6 +27,7 @@ class OpenAiV1Test {
   }
 
   @Test
+  @Disabled("Assumes not yet enables and leads to Internal Server Error (500)")
   void testCreateStreamingResponse() {
     try (final var streamResponse =
         service.createStreamingResponse("What is the capital of France?")) {
@@ -39,15 +41,6 @@ class OpenAiV1Test {
           events.stream().anyMatch(event -> event.outputTextDelta().isPresent());
       assertThat(hasTextDeltas).isTrue();
     }
-  }
-
-  @Test
-  void testCreateChatCompletion() {
-    final var response = service.createChatCompletion("What is the capital of France?");
-    assertThat(response).isNotNull();
-    assertThat(response.choices()).isNotEmpty();
-    assertThat(response.choices().get(0).message().content()).isPresent();
-    assertThat(response.choices().get(0).message().content().get()).isNotEmpty();
   }
 
   @Test
